Read commit

+++ WorkSharp/Controllers/UserProjectsController.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using AutoMapper;
 using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using WorkSharp.DAL.DbModels;
 using WorkSharp.DAL.EFCoreRepository;
@@ -11,24 +12,28 @@ using WorkSharp.ViewModels;
 
 namespace WorkSharp.Controllers
 {
+    [Authorize]
     public class UserProjectsController : Controller
     {
         private readonly IGenericRepository<DbProject> _repository;
         private IMapper _mapper;
-        public UserProjectsController(IGenericRepository<DbProject> repository, IMapper mapper)
+        private UserManager<DbUser> _userManager;
+        public UserProjectsController(IGenericRepository<DbProject> repository, IMapper mapper, UserManager<DbUser> userManager)
         {
             _repository = repository;
             _mapper = mapper;
+            _userManager = userManager;
         }
-        [Authorize]
         public IActionResult Projects()
         {
-            ViewData["Projects"] = _mapper.Map<IEnumerable<ProjectViewModel>>(_repository.GetAll());
+            var userId = _userManager.GetUserAsync(HttpContext.User).Result.Id;
+            ViewData["Projects"] = _mapper.Map<IEnumerable<ProjectViewModel>>(_repository.GetAll().Where(project => project.CreatorId.Equals(userId)));
             return View("~/Views/User/Projects/Projects.cshtml");
         }
 
         public IActionResult CreateProject(ProjectViewModel projectViewModel)
         {
+            projectViewModel.CreatorId = _userManager.GetUserAsync(HttpContext.User).Result.Id;
             _repository.Create(_mapper.Map<DbProject>(projectViewModel));
             _repository.Save();
             return RedirectToAction("Projects");
@@ -36,28 +41,70 @@ namespace WorkSharp.Controllers
 
         public IActionResult RemoveProject(Guid id)
         {
-            _repository.Delete(id);
-            _repository.Save();
-            return RedirectToAction("Projects");
+            if (CheckAccess(id))
+            {
+                _repository.Delete(id);
+                _repository.Save();
+                return RedirectToAction("Projects");
+            }
+            else
+            {
+                return View("~/Views/AccessDenied.cshtml");
+            }
         }
 
-        public IActionResult Project(Guid id)
+        public IActionResult Project(Guid projectId)
         {
-            var model = _mapper.Map<ProjectViewModel>(_repository.GetById(id));
-            return View("~/Views/User/Projects/Project.cshtml", model);
+            if (CheckAccess(projectId))
+            {
+                var model = _mapper.Map<ProjectViewModel>(_repository.GetById(projectId));
+                return View("~/Views/User/Projects/Project.cshtml", model);
+            }
+            else
+            {
+                return View("~/Views/AccessDenied.cshtml");
+            }
         }
 
         public IActionResult GetEditProject(Guid id)
         {
-            var model = _mapper.Map<ProjectViewModel>(_repository.GetById(id));
-            return View("~/Views/User/Projects/EditProject.cshtml", model);
+            if (CheckAccess(id))
+            {
+                var model = _mapper.Map<ProjectViewModel>(_repository.GetById(id));
+                return View("~/Views/User/Projects/EditProject.cshtml", model);
+            }
+            else
+            {
+                return View("~/Views/AccessDenied.cshtml");
+            }
         }
 
         public IActionResult EditProject(ProjectViewModel model)
         {
-            _repository.Update(_mapper.Map<DbProject>(model));
-            _repository.Save();
-            return RedirectToAction("Project", new {id = model.Id});
+            if (CheckAccess(model.Id))
+            {
+                _repository.Update(_mapper.Map<DbProject>(model));
+                _repository.Save();
+                return RedirectToAction("Project", new { id = model.Id });
+            }
+            else
+            {
+                return View("~/Views/AccessDenied.cshtml");
+            }
+        }
+
+        private bool CheckAccess(Guid projectId)
+        {
+            var userId = _userManager.GetUserAsync(HttpContext.User).Result.Id;
+            var projectCreatorId = _repository.GetById(projectId).CreatorId;
+            if (projectCreatorId.Equals(userId))
+            {
+                return true;
+            }
+            else
+            {
+                return false;
+            }
         }
     }
 }
